Mirrored image and timer correction
The algorithm used to display an image inverted in the y axis, probably due to the way OpenCV handles images. This has been corrected by swapping Ymin and Ymax and inverting their sign.

Also, the previous timer actually measured the time it took to launch the kernel. The new timer now measures the time it took to render the image.